+++ packages/sdrr/src/utils/build.ts
@@ -3,35 +3,17 @@ import { spawn } from "child_process"
 import { Command } from "commander"
 
 import { createRouter } from "./createRouter"
+import { excludeRouter } from "./excludeRouter"
 
-export async function build(options: Record<string, string>, command: Command) {
-    const args = command.args
+export async function build(options: Record<string, string>, { args }: Command) {
+    await excludeRouter()
 
     await createRouter()
 
     if (args.length === 0) return
 
-    if (args.at(-1)?.toLowerCase() !== "--watch" && args.at(-1)?.toLowerCase() !== "-w") {
-        spawn(args.at(0)!, args.slice(1), {
-            stdio: "inherit",
-            shell: true,
-        })
-
-        return
-    }
-
-    console.log("watch mode")
-
-    const watchPath = import.meta.resolve("./watch.js").replace(process.platform === "win32" ? /^file:\/\/\// : /^file:\/\//, "")
-
-    const child = spawn(process.execPath, [watchPath])
-
-    const child2 = spawn(args.at(0)!, args.slice(1, -1), {
+    spawn(args.at(0)!, args.slice(1), {
         stdio: "inherit",
         shell: true,
     })
-
-    child.on("close", () => child2.kill())
-
-    child2.on("close", () => child.kill())
 }