+++ packages/sdrr/src/utils/excludeRouter.ts
@@ -1,0 +1,9 @@
+import { writeVsCodeSetting } from "./writeVsCodeSetting"
+
+export function excludeRouter() {
+    return writeVsCodeSetting({
+        "files.exclude": {
+            "components/Router.tsx": true,
+        },
+    })
+}