+++ packages/soda-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-react
 
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.44.0
+
 ## 0.1.10
 
 ### Patch Changes